39 polyphase channelizer and decoder updates

+++ src/main/java/io/github/dsheirer/source/IControllableFileSource.java
@@ -40,7 +40,7 @@ public interface IControllableFileSource
 
     public long getFrameCount() throws IOException;
 
-    public int getSampleRate();
+    public double getSampleRate();
 
     public void setListener(IFrameLocationListener listener);
